Move utility classes into own directory
