+++ src/Teide/GpuExecutor.h
@@ -4,7 +4,7 @@
 #include "CommandBuffer.h"
 
 #include "Teide/BasicTypes.h"
-#include "Teide/ThreadUtils.h"
+#include "Teide/Util/ThreadUtils.h"
 #include "Teide/VulkanConfig.h"
 
 #include <function2/function2.hpp>
