+++ src/Teide/Util/StaticMap.h
